requirements.txt
fastapi~=0.109.0
pydantic~=2.6.1
httpx~=0.26.0
Pillow~=10.2.0
pytest~=7.4.4
python-dotenv~=1.0.1
requests~=2.31.0
uvicorn~=0.27.1